requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
coreapi==2.3.3
pyyaml==5.1
django-rest-swagger==2.2.0
